+++ docs/whatsnew/changelog.rst
@@ -20,6 +20,9 @@ New features
 ~~~~~~~~~~~~
 .. Add here new public features (do not delete this comment)
 
+* A new reader has been added: `read_wire` (alias `read_wdf``) to read data from
+  the .wdf format (WDF) files produced by the ReniShaw WiRe software.
+  This reader is based on the `py_wdf_reader <https://github.com/alchem0x2A/py-wdf-reader>`_ package.
 
 .. section